adds missing docs to common api config types
